init: Dynamically set CONFIG_TPM at boot
Dynamically set CONFIG_TPM at boot based on the presence of /dev/tpm0.
This allows the same board/device with and without a TPM to be handled
by a single board definition and single firmware image. It also eliminates
the need to explictly define CONFIG_TPM in board config files.

Signed-off-by: Matt DeVillier <matt.devillier@puri.sm>