PoC Will stall unless I fix things in u-root myself: https://github.com/u-root/u-root/issues/2665
Whole goal from Heads desires into tinycore would be:
- cryptsetup-reencrypt (not avail)
- thin-provisioning-tools (nice to have, but might dodge it by choosing brtfs or xfs. wyng has now support for it)
- lvm2 tools (requires proper dm mods that are kernel versioned, and not dealt with under u-root either as of now)
- gpg2 toolstack (present)
- tinycore providing system libraries into a seperate tcz (otherwise check what is packed currently under u-root.cpio under modules/u-root: this is now 2.5mb compressed by xz. That won't fly)
- u-root cannot be easily compiled by tinygo as of now: not avail under debian-11 nor debian-12....

I talked a bit on tinycore channel, tried to register to forum, now got a link to hack the registration form at:https://forum.tinycorelinux.net/index.php/topic,25947.msg168262.html#msg168262

All of this is not a really good sign for a healthy inclusion if goal is not to deduplicate work and contribute to tinycore for package updates...
Vibe for the moment is: meh.

Alternatives?
- check for nix package caches, see how complicated it would be to wrap around this...
- which goes back again into why not building Heads into NixOS, just like ownerboot is doing, but using caches of already verified reproducible binares...